Carla/backend fix (#65)
* fix listing delete

* delete console logs

+++ api-myhome/src/controllers/listing.controller.js
@@ -89,7 +89,14 @@ class ListingController {
 
   async deleteListing(req, res, next) {
     const { id } = req.params;
+  
     try {
+      const existingListing = await ListingService.getListingById(id);
+  
+      if (!existingListing) {
+        return res.status(404).json({ error: 'La publicación no existe' });
+      }
+  
       await ListingService.deleteListing(id);
       return res.status(204).json();
     } catch (err) {
@@ -97,6 +104,7 @@ class ListingController {
       next(err);
     }
   }
+  
 
   async addImages(req, res, next) {
     const { id } = req.params;
